bug fix et ajout choix activation UI

+++ doc/setup/clean.sh
@@ -7,12 +7,30 @@ then
     exit 1
 fi
 
+server_service="/etc/systemd/system/consul-server.service"
+client_service="/etc/systemd/system/consul-client.service"
+
 rm -rf /etc/consul.d
 rm -rf /var/consul
-rm -rf /etc/systemd/system/consul-client.service
-rm -f /etc/systemd/system/consul-server.service
 rm -f /usr/bin/consul-client
 rm -f /usr/bin/consul-server
+
+
+if [[ -e $server_service ]]; then
+        echo "consul-server.service trouvé. Suppression..."
+        systemctl stop consul-server.service
+        systemctl disable consul-server.service
+        rm -f /etc/systemd/system/consul-server.service
+fi
+
+if [[ -e $client_service ]]; then
+        echo "consul-client.service trouvé. Suppression..."
+        systemctl stop consul-client.service
+        systemctl disable consul-client.service
+        rm -f /etc/systemd/system/consul-client.service
+fi
+systemctl daemon-reload
+
 consul_binary=`which consul`
 if [[ ! -z $consul_binary ]]; then
 	printf "consul trouvé, suppression ? (y/n) "
